getMetaDataのリファクタリング
