+++ components/link-card/views/link-card-inner.tsx
@@ -1,5 +1,4 @@
-import { Meta } from "../../../app/posts/[slug]/page";
-import { parse } from "node-html-parser";
+import getMetaData from "../../../lib/getMetaData";
 import LinkCardError from "./link-card-error";
 
 type LinkCardInnerProps = {
@@ -40,34 +39,3 @@ const LinkCardInner = async ({
 };
 
 export default LinkCardInner;
-
-// urlのメタデータを取得
-const getMetaData = async (url: string): Promise<Meta | null> => {
-  const metaData: Meta = { url, title: "", description: "", image: "" };
-  // FIXME: Error Boundary使えるかも
-  try {
-    const res = await fetch(url);
-    const text = await res.text();
-    const doms = parse(text);
-    const metas = doms.getElementsByTagName("meta");
-
-    for (const meta of metas) {
-      const np = meta.getAttribute("property");
-      if (typeof np !== "string") continue;
-      if (np === "og:title") {
-        metaData.title = meta.getAttribute("content") ?? "";
-      }
-      if (np === "og:description") {
-        metaData.description =
-          meta.getAttribute("content")?.slice(0, 100) ?? "";
-      }
-      if (np === "og:image") {
-        metaData.image = meta.getAttribute("content") ?? "";
-      }
-    }
-  } catch (e) {
-    console.error(e);
-    return null;
-  }
-  return metaData;
-};
